security_hybrid_search/requirements.txt
opensearch-py>=2.3.0
neo4j>=5.14.0
sentence-transformers>=2.2.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
pydantic>=2.4.0
fastapi>=0.100.0
uvicorn>=0.20.0
httpx>=0.25.0